Refactor Milestone page
